pyproject.toml
[project]
name = "debateagent"
version = "0.1.0"
description = "An AI-powered debate application with multi-round arguments, automated judging."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0.0",
    "httpx>=0.24.0",
    "tenacity>=8.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "google-genai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
    "duckduckgo-search>=6.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
    "mangum>=0.17.0",
]

[project.optional-dependencies]
extras = ["tokenizers<=0.20.3,>=0.13.2"]
